move adapter outside activity
