+++ sync/sync-impl/src/main/java/com/duckduckgo/sync/impl/ui/SyncActivity.kt
@@ -78,162 +78,6 @@ interface ConnectedDeviceClickListener {
     fun onRemoveDeviceClicked(device: ConnectedDevice)
 }
 
-class SyncedDevicesAdapter constructor(private val listener: ConnectedDeviceClickListener) : RecyclerView.Adapter<ViewHolder>() {
-
-    private var syncedDevices = mutableListOf<SyncDeviceListItem>()
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int,
-    ): ViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        return when (viewType) {
-            CONNECTED_DEVICE -> SyncedDeviceViewHolder(
-                inflater,
-                ItemSyncDeviceBinding.inflate(inflater, parent, false),
-                listener,
-            )
-            LOADING_ITEM -> LoadingViewHolder(ItemSyncDeviceLoadingBinding.inflate(inflater, parent, false))
-            else -> {
-                // This should never happen
-                return SyncedDeviceViewHolder(
-                    inflater,
-                    ItemSyncDeviceBinding.inflate(inflater, parent, false),
-                    listener,
-                )
-            }
-        }
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        return when (val device = syncedDevices[position]) {
-            is SyncedDevice -> {
-                if (device.loading) {
-                    LOADING_ITEM
-                } else {
-                    CONNECTED_DEVICE
-                }
-            }
-            is SyncDeviceListItem.LoadingItem -> LOADING_ITEM
-        }
-    }
-
-    override fun getItemCount(): Int {
-        return this.syncedDevices.size
-    }
-
-    override fun onBindViewHolder(
-        holder: ViewHolder,
-        position: Int,
-    ) {
-        when (holder) {
-            is SyncedDeviceViewHolder -> {
-                holder.bind(syncedDevices[position] as SyncedDevice)
-            }
-
-            is LoadingViewHolder -> {
-                holder.bind()
-            }
-        }
-    }
-
-    @SuppressLint("NotifyDataSetChanged")
-    fun updateData(data: List<SyncDeviceListItem>) {
-        this.syncedDevices.clear().also { this.syncedDevices.addAll(data) }
-        notifyDataSetChanged() // there's a weird bug when using DiffUtil with duplicated devices after login, need to investigate
-    }
-
-    private class DiffCallback(
-        private val old: List<SyncDeviceListItem>,
-        private val new: List<SyncDeviceListItem>,
-    ) :
-        DiffUtil.Callback() {
-        override fun getOldListSize() = old.size
-
-        override fun getNewListSize() = new.size
-
-        override fun areItemsTheSame(
-            oldItemPosition: Int,
-            newItemPosition: Int,
-        ): Boolean {
-            if (old[oldItemPosition] is SyncedDevice && new[newItemPosition] is SyncedDevice) {
-                return (old[oldItemPosition] as SyncedDevice).device.deviceId == (new[newItemPosition] as SyncedDevice).device.deviceId
-            }
-            return old[oldItemPosition] == new[newItemPosition]
-        }
-
-        override fun areContentsTheSame(
-            oldItemPosition: Int,
-            newItemPosition: Int,
-        ): Boolean {
-            return old[oldItemPosition] == new[newItemPosition]
-        }
-    }
-
-    companion object {
-        private const val CONNECTED_DEVICE = 0
-        private const val LOADING_ITEM = 1
-    }
-}
-
-class SyncedDeviceViewHolder(
-    private val layoutInflater: LayoutInflater,
-    private val binding: ItemSyncDeviceBinding,
-    private val listener: ConnectedDeviceClickListener,
-) : ViewHolder(binding.root) {
-    private val context = binding.root.context
-    fun bind(syncDevice: SyncedDevice) {
-        with(syncDevice.device) {
-            if (thisDevice) {
-                binding.localSyncDevice.show()
-                binding.localSyncDevice.setLeadingIcon(deviceType.type().asDrawableRes())
-                binding.localSyncDevice.setPrimaryText(deviceName)
-                binding.localSyncDevice.setSecondaryText(context.getString(R.string.sync_device_this_device_hint))
-                binding.localSyncDevice.setTrailingIconClickListener {
-                    showLocalOverFlowMenu(it, syncDevice)
-                }
-            } else {
-                binding.remoteSyncDevice.show()
-                binding.remoteSyncDevice.setLeadingIcon(deviceType.type().asDrawableRes())
-                binding.remoteSyncDevice.setPrimaryText(deviceName)
-                binding.remoteSyncDevice.setTrailingIconClickListener {
-                    showRemoteOverFlowMenu(it, syncDevice)
-                }
-            }
-        }
-    }
-
-    private fun showLocalOverFlowMenu(
-        anchor: View,
-        syncDevice: SyncedDevice,
-    ) {
-        val popupMenu = PopupMenu(layoutInflater, R.layout.popup_windows_edit_device_menu)
-        val view = popupMenu.contentView
-        popupMenu.apply {
-            onMenuItemClicked(view.findViewById(R.id.edit)) { listener.onEditDeviceClicked(syncDevice.device) }
-        }
-        popupMenu.show(binding.root, anchor)
-    }
-
-    private fun showRemoteOverFlowMenu(
-        anchor: View,
-        syncDevice: SyncedDevice,
-    ) {
-        val popupMenu = PopupMenu(layoutInflater, R.layout.popup_windows_remove_device_menu)
-        val view = popupMenu.contentView
-        popupMenu.apply {
-            onMenuItemClicked(view.findViewById(R.id.remove)) { listener.onRemoveDeviceClicked(syncDevice.device) }
-        }
-        popupMenu.show(binding.root, anchor)
-    }
-}
-
-class LoadingViewHolder(val binding: ItemSyncDeviceLoadingBinding) : ViewHolder(binding.root) {
-    fun bind() {
-        binding.shimmerFrameLayout.startShimmer()
-    }
-}
-
 @InjectWith(ActivityScope::class)
 @ContributeToActivityStarter(SyncActivityWithEmptyParams::class)
 class SyncActivity : DuckDuckGoActivity() {
